Safari-609.1.11.1/LayoutTests/http/wpt/workers/promise-unhandled-rejection.any-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Reject

Harness Error (FAIL), message = undefined

PASS UnhandledRejection event occurs if a rejected promise is not handled. 
